Remove suffix "storage" from module names that have it
Will result in a more consistent naming style for Python modules and
backends. The old names are still available as aliases, but produce a
warning that they will be removed in django-storages 2.0.
